rework Page component
