+++ query-connector/src/app/(pages)/landingPage/landpingPage.module.scss
@@ -1,0 +1,14 @@
+.pageSubtitle {
+    margin: 2rem 0 0.5rem 0;
+}
+
+.pageSubtitle:last-of-type {
+    margin-top: 1.5rem;
+}
+
+.pageContent {
+    font-weight: 300;
+    margin-top: 0;
+    font-size: 1rem;
+    margin-bottom: 1rem;
+}
