+++ query-connector/src/app/(pages)/queryBuilding/page.tsx
@@ -6,8 +6,8 @@ import "react-toastify/dist/ReactToastify.css";
 import { useState } from "react";
 import { EMPTY_QUERY_SELECTION } from "./utils";
 import BuildFromTemplates from "./buildFromTemplates/BuildFromTemplates";
-import { ToastContainer } from "react-toastify";
-
+import { useContext, useEffect } from "react";
+import { DataContext } from "@/app/shared/DataProvider";
 /**
  * Component for Query Building Flow
  * @returns The Query Building component flow
@@ -17,15 +17,23 @@ const QueryBuilding: React.FC = () => {
     structuredClone(EMPTY_QUERY_SELECTION),
   );
   const [buildStep, setBuildStep] = useState<BuildStep>("selection");
+  const ctx = useContext(DataContext);
+
+  // update the current page details when switching between build steps
+  useEffect(() => {
+    ctx?.setCurrentPage(buildStep);
+  }, [buildStep]);
+
+  useEffect(() => {
+    ctx?.setToastConfig({
+      position: "bottom-left",
+      stacked: true,
+      hideProgressBar: true,
+    });
+  }, []);
 
   return (
     <>
-      <ToastContainer
-        position="bottom-left"
-        icon={false}
-        stacked
-        hideProgressBar
-      />
       {buildStep === "selection" && (
         <QuerySelection
           selectedQuery={selectedQuery}
