+++ query-connector/src/app/ui/styles/_variables.scss
@@ -174,3 +174,4 @@ $checkbox-border: $gray-500;
 $header-content-height: 3.5rem;
 $footer-content-height: 5rem;
 $default-table-padding: 1.5rem;
+$site-alert-height: 2.5rem;
