+++ query-connector/src/app/ui/styles/custom-styles.scss
@@ -37,6 +37,7 @@ h1 {
 
 .main-body {
   background-color: rgba(247, 249, 250, 1);
+  display: flex;
   flex: 1 1 auto;
 }
 
@@ -128,11 +129,7 @@ h1 {
 }
 
 .gradient-blue-background {
-  background: linear-gradient(
-    109.54deg,
-    rgba(173, 207, 220, 0.025) 8.02%,
-    rgba(46, 98, 118, 0.5) 117.53%
-  );
+  background: linear-gradient(109.54deg, rgba(173, 207, 220, 0.025) 8.02%, rgba(46, 98, 118, 0.5) 117.53%);
 
   padding: 0.5rem;
   /* Optional: some padding to make the background more noticeable */
@@ -327,19 +324,11 @@ hr.custom-hr {
   font-weight: bold;
 }
 
-ul.usa-sidenav
-  > li.usa-sidenav__item
-  > ul.usa-sidenav__sublist
-  > li.usa-sidenav__item
-  > a {
+ul.usa-sidenav > li.usa-sidenav__item > ul.usa-sidenav__sublist > li.usa-sidenav__item > a {
   font-weight: bold;
 }
 
-ul.usa-sidenav
-  > li.usa-sidenav__item
-  > ul.usa-sidenav__sublist
-  > li.usa-sidenav__item
-  > a.usa-current {
+ul.usa-sidenav > li.usa-sidenav__item > ul.usa-sidenav__sublist > li.usa-sidenav__item > a.usa-current {
   &::after {
     background-color: rgb(0, 94, 162);
     content: "";
@@ -517,13 +506,12 @@ ul.usa-sidenav
 
 .custom-alert {
   margin-top: 0;
+  height: $site-alert-height;
 }
 
 .usa-alert__text {
   font-size: 0.875rem;
 }
-
-
 
 #fhir-server-modal {
   .usa-modal__close {
@@ -545,7 +533,8 @@ ul.usa-sidenav
     .usa-modal__main {
       max-width: 34rem;
 
-      input, select {
+      input,
+      select {
         max-width: 100%;
       }
     }
@@ -564,7 +553,7 @@ ul.usa-sidenav
 
     #modal-cancel-button {
       background-color: transparent;
-      
+
       &:hover {
         box-shadow: none;
       }
